requirements.txt
discord.py
praw
python-dotenv
flask
beautifulsoup4
requests